[Refactor] 클리어, 실패 함수 위치 변경
스테이지 클리어, 스테이지 실패 함수의 위치를 스테이지매니저로 옮겼습니다.
